feat: remove presentation layer from game
